Extract server apps and funcs from `sdk` into `server` (#388)
This separates server functionalities (Repository apps, in the future 
also Discovery, and Registries) from the core SDK to improve modularity 
and maintainability. It also includes some refactoring to avoid code 
duplication in future.

Changes:
- Moved all server-related classes and functions from `sdk` to `server`
- Splitted `http.py` into `repository.py`, `base.py` and 
  `util/converters.py`
- Created parent classes `BaseWSGIApp` and `ObjectStoreWSGIApp` for 
  repository app class, which will also be used for discovery and 
  registry apps in the future
- Added an initial pyproject.toml for the server app
- Refactored  `object_hook()` by extracting the mapping dict into 
  `_get_aas_class_parsers()`
- Refactored  `default()` by extracting the mapping dict into 
  `_get_aas_class_serializers()`
- Refactored `_create_dict()`
- Created a JSON_AAS_TOP_LEVEL_KEYS_TO_TYPES tuple with top-level JSON 
  keys and the corresponding SDK types to reuse it in JSON 
  de-/serialization methods
