+++ .github/workflows/ci.yml
@@ -254,39 +254,10 @@ jobs:
         pip install .[dev]
     - name: Check typing with MyPy
       run: |
-        mypy ./aas_compliance_tool test
+        mypy aas_compliance_tool test
     - name: Check code style with PyCodestyle
       run: |
-        pycodestyle --count --max-line-length 120 ./aas_compliance_tool test
-
-  compliance-tool-readme-codeblocks:
-    # This job runs the same static code analysis (mypy and pycodestyle) on the codeblocks in our docstrings.
-    runs-on: ubuntu-latest
-
-    defaults:
-      run:
-        working-directory: ./compliance_tool
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python ${{ env.X_PYTHON_MIN_VERSION }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ env.X_PYTHON_MIN_VERSION }}
-    - name: Install Python dependencies
-      # install the local sdk in editable mode so it does not get overwritten
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e ../sdk[dev]
-        pip install .[dev]
-    - name: Check typing with MyPy
-      run: |
-        mypy <(codeblocks python README.md)
-    - name: Check code style with PyCodestyle
-      run: |
-        codeblocks --wrap python README.md | pycodestyle --count --max-line-length 120 -
-    - name: Run readme codeblocks with Python
-      run: |
-        codeblocks python README.md | python
+        pycodestyle --count --max-line-length 120 aas_compliance_tool test
 
   compliance-tool-package:
     # This job checks if we can build our compliance_tool package
@@ -308,6 +279,36 @@ jobs:
     - name: Create source and wheel dist
       run: |
         python -m build
+
+ #server-test:
+    # TODO: This job runs the unittests on the python versions specified down at the matrix
+    #   and aas-test-engines on the server
+
+
+  server-static-analysis:
+    # This job runs static code analysis, namely pycodestyle and mypy
+    runs-on: ubuntu-latest
+
+    defaults:
+      run:
+        working-directory: ./server/app
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python ${{ env.X_PYTHON_MIN_VERSION }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ env.X_PYTHON_MIN_VERSION }}
+    - name: Install Python dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install ../../sdk
+        pip install .[dev]
+    - name: Check typing with MyPy
+      run: |
+        mypy .
+    - name: Check code style with PyCodestyle
+      run: |
+        pycodestyle --count --max-line-length 120 .
 
   server-package:
     # This job checks if we can build our server package
