+++ .gitignore
@@ -29,6 +29,7 @@ sdk/test/adapter/schemas
 # Ignore dynamically generated version file
 sdk/basyx/version.py
 compliance_tool/aas_compliance_tool/version.py
+server/app/version.py
 
 # ignore the content of the server storage
 server/storage/
