+++ server/app/main.py
@@ -6,7 +6,7 @@ from basyx.aas import model, adapter
 from basyx.aas.adapter import aasx
 
 from basyx.aas.backend.local_file import LocalFileObjectStore
-from basyx.aas.adapter.http import WSGIApp
+from interfaces.repository import WSGIApp
 
 storage_path = os.getenv("STORAGE_PATH", "/storage")
 storage_type = os.getenv("STORAGE_TYPE", "LOCAL_FILE_READ_ONLY")
