+++ server/test/interfaces/test_repository.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024 the Eclipse BaSyx Authors
+# Copyright (c) 2025 the Eclipse BaSyx Authors
 #
 # This program and the accompanying materials are made available under the terms of the MIT License, available in
 # the LICENSE file of this project.
@@ -34,7 +34,7 @@ import hypothesis.strategies
 
 from basyx.aas import model
 from basyx.aas.adapter.aasx import DictSupplementaryFileContainer
-from basyx.aas.adapter.http import WSGIApp
+from server.app.interfaces.repository import WSGIApp
 from basyx.aas.examples.data.example_aas import create_full_example
 
 from typing import Set
